Fix a Few CodeQL Warnings

+++ src/perf/lib/SecNetPerfMain.cpp
@@ -155,8 +155,8 @@ QuicMainStart(
     if ((CpuStr = GetValue(argc, argv, "cpu")) != nullptr) {
         SetConfig = true;
         if (strtol(CpuStr, nullptr, 10) == -1) {
-            for (uint16_t i = 0; i < CxPlatProcCount() && Config->ProcessorCount < 256; ++i) {
-                Config->ProcessorList[Config->ProcessorCount++] = i;
+            for (uint32_t i = 0; i < CxPlatProcCount() && Config->ProcessorCount < 256; ++i) {
+                Config->ProcessorList[Config->ProcessorCount++] = (uint16_t)i;
             }
         } else {
             do {
